2001-07-05  [ Ela ]  <ela@lkcc.org>
        * README: Added a paragraph about build problems due to
        dynamic linking.

2001-07-05  [ Ela ]  <ela@lkcc.org>

        * core.c (svz_sendfile): Partly figured out how sendfile()
        works on OSF4.0g (True64).

+++ src/fakeident-server/ident-proto.c
@@ -19,7 +19,7 @@
  * the Free Software Foundation, Inc., 59 Temple Place - Suite 330,
  * Boston, MA 02111-1307, USA.  
  *
- * $Id: ident-proto.c,v 1.13 2001/07/04 20:14:56 raimi Exp $
+ * $Id: ident-proto.c,v 1.14 2001/07/05 17:29:17 ela Exp $
  *
  */
 
@@ -149,7 +149,7 @@ fakeident_detect_proto (svz_server_t *server, svz_socket_t *sock)
     goto out;
 
   /* number */
-  for (; p < end && isdigit (*p); p++);
+  for (; p < end && isdigit ((int) *p); p++);
   if (p == end)
     goto out;
